+++ hook_event_dispatcher.module
@@ -7,7 +7,6 @@
 
 use Drupal\Core\Session\AccountInterface;
 use Drupal\Core\Url;
-use Drupal\hook_event_dispatcher\Event\Cron\CronEvent;
 use Drupal\hook_event_dispatcher\Event\Language\LanguageSwitchLinksAlterEvent;
 use Drupal\hook_event_dispatcher\Event\Path\PathDeleteEvent;
 use Drupal\hook_event_dispatcher\Event\Path\PathInsertEvent;
@@ -298,18 +297,6 @@ function hook_event_dispatcher_toolbar_alter(array &$items) {
   /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
   $manager = Drupal::service('hook_event_dispatcher.manager');
   $event = new ToolbarAlterEvent($items);
-  $manager->register($event);
-}
-
-/**
- * Implements hook_cron().
- *
- * {@inheritdoc}
- */
-function hook_event_dispatcher_cron() {
-  /* @var \Drupal\hook_event_dispatcher\Manager\HookEventDispatcherManagerInterface $manager */
-  $manager = Drupal::service('hook_event_dispatcher.manager');
-  $event = new CronEvent();
   $manager->register($event);
 }
